chore(v2): restructure folder
