+++ packages/docusaurus/lib/client/App.js
