+++ packages/docusaurus/lib/client/clientEntry.js
