+++ packages/docusaurus/lib/client/exports/Head.js
