+++ packages/docusaurus/lib/client/serverEntry.js
