+++ packages/docusaurus/lib/client/templates/index.html.template.ejs
