+++ packages/docusaurus/lib/client/templates/metadata.template.ejs
