+++ packages/docusaurus/lib/client/templates/ssr.html.template.js
