+++ packages/docusaurus/lib/commands/build.js
@@ -13,7 +13,7 @@ const path = require('path');
 const chalk = require('chalk');
 const fs = require('fs-extra');
 const globby = require('globby');
-const load = require('../load');
+const load = require('../server/load');
 const createServerConfig = require('../webpack/server');
 const createClientConfig = require('../webpack/client');
 const {applyConfigureWebpack} = require('../webpack/utils');
