+++ packages/docusaurus/lib/commands/eject.js
@@ -10,7 +10,7 @@ const chalk = require('chalk');
 const path = require('path');
 
 module.exports = async function eject(siteDir) {
-  const defaultTheme = path.resolve(__dirname, '..', 'theme');
+  const defaultTheme = path.resolve(__dirname, '..', 'default-theme');
   const customTheme = path.resolve(siteDir, 'theme');
   await fs.copy(defaultTheme, customTheme);
