+++ packages/docusaurus/lib/default-theme/Doc/index.js
