+++ packages/docusaurus/lib/default-theme/DocsPaginator/index.js
