+++ packages/docusaurus/lib/default-theme/Footer/index.js
