+++ packages/docusaurus/lib/default-theme/Layout/index.js
