+++ packages/docusaurus/lib/default-theme/Navbar/index.js
