+++ packages/docusaurus/lib/default-theme/Navbar/styles.module.css
