+++ packages/docusaurus/lib/default-theme/Search/index.js
