+++ packages/docusaurus/lib/default-theme/Search/styles.css
