+++ packages/docusaurus/lib/default-theme/Sidebar/index.js
