+++ packages/docusaurus/lib/default-theme/Sidebar/styles.module.css
