+++ packages/docusaurus/lib/server/load/plugins.js
