+++ packages/docusaurus/lib/server/load/routes.js
